+++ workspace/node/bookmarks/src/parsers/index.ts
@@ -1,1 +1,3 @@
+export * from "./netscape";
+export * from "./parser";
 export * from "./xbel";